PTI-385: Rename integration file and variable names.
