+++ api/src/integrations/epic/certificates-amendment-utils.js
@@ -1,37 +1,37 @@
 'use strict';
 
 const mongoose = require('mongoose');
-const defaultLog = require('../../utils/logger')('epic-orders');
+const defaultLog = require('../../utils/logger')('epic-amendments');
 const RECORD_TYPE = require('../../utils/constants/record-type-enum');
-const { preTransformRecord } = require('./epic-utils');
+const EpicUtils = require('./epic-utils');
 
 /**
- * Epic Order record handler for { type: 'Order', milestone: 'Other' }.
+ * Epic Amendment record handler for { type: 'Amendment Package', milestone: 'Amendment' }.
  *
  * Must contain the following functions:
- * - transformRecord: (object) => Order
- * - saveRecord: (Order) => any
+ * - transformRecord: (object) => Certificate
+ * - saveRecord: (Certificate) => any
  *
- * @class EpicOrdersOther
+ * @class CertificatesAmendment
  */
-class EpicOrdersOther {
+class CertificatesAmendment {
   /**
-   * Creates an instance of EpicOrdersOther.
+   * Creates an instance of CertificatesAmendment.
    *
    * @param {*} auth_payload user information for auditing
-   * @memberof EpicOrdersOther
+   * @memberof CertificatesAmendment
    */
   constructor(auth_payload) {
     this.auth_payload = auth_payload;
   }
 
   /**
-   * Transform an Epic order record into a NRPTI Order record.
+   * Transform an Epic amendment record into a NRPTI Certificate record.
    *
-   * @param {object} epicRecord Epic order record (required)
-   * @returns {Order} NRPTI order record.
+   * @param {object} epicRecord Epic amendment record (required)
+   * @returns {Certificate} NRPTI certificate record.
    * @throws {Error} if record is not provided.
-   * @memberof EpicOrdersOther
+   * @memberof CertificatesAmendment
    */
   async transformRecord(epicRecord) {
     if (!epicRecord) {
@@ -39,10 +39,21 @@ class EpicOrdersOther {
     }
 
     // Apply common Epic pre-processing/transformations
-    epicRecord = preTransformRecord(epicRecord);
+    epicRecord = EpicUtils.preTransformRecord(epicRecord);
+
+    // Generate a link that will get us the document when placed in an href.
+    const attachments = [];
+    if (epicRecord._id && epicRecord.documentFileName) {
+      attachments.push({
+        url: `https://projects.eao.gov.bc.ca/api/document/${epicRecord._id}/fetch/${encodeURIComponent(
+          epicRecord.documentFileName
+        )}`,
+        fileName: epicRecord.documentFileName
+      });
+    }
 
     return {
-      _schemaName: RECORD_TYPE.Order._schemaName,
+      _schemaName: RECORD_TYPE.Certificate._schemaName,
       _epicProjectId: (epicRecord.project && epicRecord.project._id) || '',
       _sourceRefId: epicRecord._id || '',
       _epicMilestoneId: epicRecord.milestone || '',
@@ -51,7 +62,8 @@ class EpicOrdersOther {
       write: ['sysadmin'],
 
       recordName: epicRecord.displayName || '',
-      recordType: RECORD_TYPE.Order.displayName,
+      recordType: RECORD_TYPE.Certificate.displayName,
+      recordSubtype: 'Amendment',
       dateIssued: epicRecord.documentDate || null,
       issuingAgency: 'Environmental Assessment Office',
       author: epicRecord.documentAuthor || '',
@@ -59,6 +71,7 @@ class EpicOrdersOther {
       projectName: (epicRecord.project && epicRecord.project.name) || '',
       location: (epicRecord.project && epicRecord.project.location) || '',
       centroid: (epicRecord.project && epicRecord.project.centroid) || '',
+      attachments: attachments,
 
       dateAdded: new Date(),
       dateUpdated: new Date(),
@@ -70,33 +83,32 @@ class EpicOrdersOther {
   }
 
   /**
-   * Persist a NRPTI Order record to the database.
+   * Persist a NRPTI Certificate record to the database.
    *
    * @async
-   * @param {Order} orderRecord NRPTI Order record (required)
+   * @param {Certificate} certificateRecord NRPTI Certificate record (required)
    * @returns {string} status of the add/update operations.
-   * @memberof EpicOrdersOther
+   * @memberof CertificatesAmendment
    */
-  async saveRecord(orderRecord) {
-    if (!orderRecord) {
+  async saveRecord(certificateRecord) {
+    if (!certificateRecord) {
       throw Error('saveRecord - required record must be non-null.');
     }
 
     try {
-      const Order = mongoose.model(RECORD_TYPE.Order._schemaName);
+      const Certificate = mongoose.model(RECORD_TYPE.Certificate._schemaName);
 
-      const record = await Order.findOneAndUpdate(
-        { _schemaName: RECORD_TYPE.Order._schemaName, _sourceRefId: orderRecord._sourceRefId },
-        { $set: orderRecord },
+      const record = await Certificate.findOneAndUpdate(
+        { _schemaName: RECORD_TYPE.Certificate._schemaName, _sourceRefId: certificateRecord._sourceRefId },
+        { $set: certificateRecord },
         { upsert: true, new: true }
       );
 
       return record;
     } catch (error) {
-      defaultLog.error(`Failed to save Epic Order record: ${error.message}`);
-      defaultLog.debug(`Failed to save Epic Order record - error.stack: ${error.stack}`);
+      defaultLog.error(`Failed to save Epic Certificate record: ${error.message}`);
     }
   }
 }
 
-module.exports = EpicOrdersOther;
+module.exports = CertificatesAmendment;
