+++ api/src/integrations/epic/orders-utils.js
@@ -1,37 +1,37 @@
 'use strict';
 
 const mongoose = require('mongoose');
-const defaultLog = require('../../utils/logger')('epic-amendments');
+const defaultLog = require('../../utils/logger')('epic-orders');
 const RECORD_TYPE = require('../../utils/constants/record-type-enum');
 const EpicUtils = require('./epic-utils');
 
 /**
- * Epic Amendment record handler for { type: 'Amendment Package', milestone: 'Amendment' }.
+ * Epic Order record handler for { type: 'Order', milestone: 'Compliance and Enforcement' }.
  *
  * Must contain the following functions:
- * - transformRecord: (object) => Certificate
- * - saveRecord: (Certificate) => any
+ * - transformRecord: (object) => Order
+ * - saveRecord: (Order) => any
  *
- * @class EpicCertificatesAmendment
+ * @class Orders
  */
-class EpicCertificatesAmendment {
+class Orders {
   /**
-   * Creates an instance of EpicCertificatesAmendment.
+   * Creates an instance of Orders.
    *
    * @param {*} auth_payload user information for auditing
-   * @memberof EpicCertificatesAmendment
+   * @memberof Orders
    */
   constructor(auth_payload) {
     this.auth_payload = auth_payload;
   }
 
   /**
-   * Transform an Epic amendment record into a NRPTI Certificate record.
+   * Transform an Epic order record into a NRPTI Order record.
    *
-   * @param {object} epicRecord Epic amendment record (required)
-   * @returns {Certificate} NRPTI certificate record.
+   * @param {object} epicRecord Epic order record (required)
+   * @returns {Order} NRPTI order record.
    * @throws {Error} if record is not provided.
-   * @memberof EpicCertificatesAmendment
+   * @memberof Orders
    */
   async transformRecord(epicRecord) {
     if (!epicRecord) {
@@ -53,7 +53,7 @@ class EpicCertificatesAmendment {
     }
 
     return {
-      _schemaName: RECORD_TYPE.Certificate._schemaName,
+      _schemaName: RECORD_TYPE.Order._schemaName,
       _epicProjectId: (epicRecord.project && epicRecord.project._id) || '',
       _sourceRefId: epicRecord._id || '',
       _epicMilestoneId: epicRecord.milestone || '',
@@ -62,15 +62,18 @@ class EpicCertificatesAmendment {
       write: ['sysadmin'],
 
       recordName: epicRecord.displayName || '',
-      recordType: RECORD_TYPE.Certificate.displayName,
-      recordSubtype: 'Amendment',
-      dateIssued: epicRecord.documentDate || null,
+      recordType: RECORD_TYPE.Order.displayName,
+      // recordSubtype: // No mapping
+      dateIssued: epicRecord.datePosted || null,
       issuingAgency: 'Environmental Assessment Office',
       author: epicRecord.documentAuthor || '',
       legislation: (epicRecord.project && epicRecord.project.legislation) || '',
+      // issuedTo: // No mapping
       projectName: (epicRecord.project && epicRecord.project.name) || '',
       location: (epicRecord.project && epicRecord.project.location) || '',
       centroid: (epicRecord.project && epicRecord.project.centroid) || '',
+      // outcomeStatus: // No mapping
+      // outcomeDescription: // No mapping
       attachments: attachments,
 
       dateAdded: new Date(),
@@ -83,32 +86,32 @@ class EpicCertificatesAmendment {
   }
 
   /**
-   * Persist a NRPTI Certificate record to the database.
+   * Persist a NRPTI Order record to the database.
    *
    * @async
-   * @param {Certificate} certificateRecord NRPTI Certificate record (required)
+   * @param {Order} orderRecord NRPTI Order record (required)
    * @returns {string} status of the add/update operations.
-   * @memberof EpicCertificatesAmendment
+   * @memberof Orders
    */
-  async saveRecord(certificateRecord) {
-    if (!certificateRecord) {
+  async saveRecord(orderRecord) {
+    if (!orderRecord) {
       throw Error('saveRecord - required record must be non-null.');
     }
 
     try {
-      const Certificate = mongoose.model(RECORD_TYPE.Certificate._schemaName);
+      const Order = mongoose.model(RECORD_TYPE.Order._schemaName);
 
-      const record = await Certificate.findOneAndUpdate(
-        { _schemaName: RECORD_TYPE.Certificate._schemaName, _sourceRefId: certificateRecord._sourceRefId },
-        { $set: certificateRecord },
+      const record = await Order.findOneAndUpdate(
+        { _schemaName: RECORD_TYPE.Order._schemaName, _sourceRefId: orderRecord._sourceRefId },
+        { $set: orderRecord },
         { upsert: true, new: true }
       );
 
       return record;
     } catch (error) {
-      defaultLog.error(`Failed to save Epic Certificate record: ${error.message}`);
+      defaultLog.error(`Failed to save Epic Order record: ${error.message}`);
     }
   }
 }
 
-module.exports = EpicCertificatesAmendment;
+module.exports = Orders;
