+++ api/src/integrations/epic/orders-utils.test.js
@@ -1,27 +1,27 @@
-const EpicInspections = require('./epic-inspections');
+const OrdersUtils = require('./orders-utils');
 const RECORD_TYPE = require('../../utils/constants/record-type-enum');
 
-describe('EpicInspections', () => {
+describe('OrdersUtils', () => {
   describe('transformRecord', () => {
     it('throws error if no epicRecord provided', async () => {
-      const epicInspections = new EpicInspections();
+      const ordersUtils = new OrdersUtils();
 
       try {
-        await epicInspections.transformRecord();
+        await ordersUtils.transformRecord();
       } catch (error) {
         expect(error).toEqual(new Error('transformRecord - required record must be non-null.'));
       }
     });
 
     it('returns a default nrpti record when empty epicRecord provided', async () => {
-      const epicInspections = new EpicInspections();
+      const ordersUtils = new OrdersUtils();
 
       const epicRecord = {};
 
-      const actualRecord = await epicInspections.transformRecord(epicRecord);
+      const actualRecord = await ordersUtils.transformRecord(epicRecord);
 
       const expectedRecord = {
-        _schemaName: RECORD_TYPE.Inspection._schemaName,
+        _schemaName: RECORD_TYPE.Order._schemaName,
 
         _epicProjectId: '',
         _sourceRefId: '',
@@ -31,7 +31,7 @@ describe('EpicInspections', () => {
         write: ['sysadmin'],
 
         recordName: '',
-        recordType: RECORD_TYPE.Inspection.displayName,
+        recordType: RECORD_TYPE.Order.displayName,
         dateIssued: null,
         issuingAgency: 'Environmental Assessment Office',
         author: '',
@@ -52,7 +52,7 @@ describe('EpicInspections', () => {
     });
 
     it('returns a nrpti record with all supported epicRecord fields populated', async () => {
-      const epicInspections = new EpicInspections();
+      const ordersUtils = new OrdersUtils();
 
       const epicRecord = {
         _id: 123,
@@ -65,10 +65,10 @@ describe('EpicInspections', () => {
         milestone: 'milestone'
       };
 
-      const actualRecord = await epicInspections.transformRecord(epicRecord);
+      const actualRecord = await ordersUtils.transformRecord(epicRecord);
 
       const expectedRecord = {
-        _schemaName: RECORD_TYPE.Inspection._schemaName,
+        _schemaName: RECORD_TYPE.Order._schemaName,
 
         _epicProjectId: '',
         _sourceRefId: 123,
@@ -78,7 +78,7 @@ describe('EpicInspections', () => {
         write: ['sysadmin'],
 
         recordName: 'docDisplay',
-        recordType: RECORD_TYPE.Inspection.displayName,
+        recordType: RECORD_TYPE.Order.displayName,
         dateIssued: null,
         issuingAgency: 'Environmental Assessment Office',
         author: '',
@@ -100,14 +100,14 @@ describe('EpicInspections', () => {
   });
 
   describe('saveRecord', () => {
-    it('throws error if no inspection record provided', async () => {
-      const epicInspections = new EpicInspections();
-      await expect(epicInspections.saveRecord()).rejects.toThrow(
+    it('throws error if no order record provided', async () => {
+      const ordersUtils = new OrdersUtils();
+      await expect(ordersUtils.saveRecord()).rejects.toThrow(
         new Error('saveRecord - required record must be non-null.')
       );
     });
 
-    it('catches any errors thrown when creating/saving the inspection record', async () => {
+    it('catches any errors thrown when creating/saving the order record', async () => {
       // create mock save function
       const mockFindOneAndUpdate = jest.fn(() => {
         throw Error('this should not be thrown');
@@ -119,14 +119,14 @@ describe('EpicInspections', () => {
         return { findOneAndUpdate: mockFindOneAndUpdate };
       });
 
-      const epicInspections = new EpicInspections();
+      const ordersUtils = new OrdersUtils();
 
-      const inspectionRecord = { _id: '321' };
+      const orderRecord = { _id: '321' };
 
-      await expect(epicInspections.saveRecord(inspectionRecord)).resolves.not.toThrow();
+      await expect(ordersUtils.saveRecord(orderRecord)).resolves.not.toThrow();
     });
 
-    it('creates and saves a new inspection record', async () => {
+    it('creates and saves a new order record', async () => {
       // create mock save function
       const mockFindOneAndUpdate = jest.fn(() => Promise.resolve('saved!'));
 
@@ -136,11 +136,11 @@ describe('EpicInspections', () => {
         return { findOneAndUpdate: mockFindOneAndUpdate };
       });
 
-      const epicInspections = new EpicInspections();
+      const ordersUtils = new OrdersUtils();
 
-      const inspectionRecord = { _id: '123' };
+      const orderRecord = { _id: '123' };
 
-      const dbStatus = await epicInspections.saveRecord(inspectionRecord);
+      const dbStatus = await ordersUtils.saveRecord(orderRecord);
 
       expect(mockFindOneAndUpdate).toHaveBeenCalledTimes(1);
       expect(dbStatus).toEqual('saved!');
